gathering server/requirements.txt
Bottleneck==1.3.5
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==39.0.2
et-xmlfile==1.1.0
greenlet==2.0.2
idna==3.4
jwt==1.3.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numexpr==2.8.4
numpy==1.23.5
openpyxl==3.1.1
packaging==22.0
pandas==1.5.3
pip==23.0.1
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7
pyupbit==0.2.33
requests==2.28.2
schedule==1.1.0
setuptools==65.6.3
six==1.16.0
SQLAlchemy==2.0.5.post1
TA-Lib==0.4.24
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.14
websockets==10.4
wheel==0.38.4
wincertstore==0.2
